Unify pages, actions and reducers
