+++ src/actions/room-booking-actions.js
@@ -51,10 +51,7 @@ export const ROOM_BOOKING_REFUNDED = 'ROOM_BOOKING_REFUNDED';
 
 export const ROOM_BOOKING_CANCELED = 'ROOM_BOOKING_CANCELED';
 
-export const OFFLINE_ROOM_BOOKING_ADDED = 'OFFLINE_ROOM_BOOKING_ADDED';
-export const RECEIVE_OFFLINE_ROOM_BOOKING_AVAILABILITY = 'RECEIVE_OFFLINE_ROOM_BOOKING_AVAILABILITY';
-export const UPDATE_ROOM_BOOKING_RESERVATION = 'UPDATE_ROOM_BOOKING_RESERVATION';
-export const ROOM_BOOKING_RESERVATION_UPDATED = 'ROOM_BOOKING_RESERVATION_UPDATED';
+export const RECEIVE_ROOM_BOOKING_AVAILABILITY = 'RECEIVE_ROOM_BOOKING_AVAILABILITY';
 
 export const getRoomBookings = (term = null, page = 1, perPage = 10, order = 'start_datetime', orderDir = 1) => async (dispatch, getState) => {
 
@@ -174,10 +171,10 @@ export const saveRoomBooking = (entity) => async (dispatch, getState) => {
 
     if (entity.id) {
 
-        putRequest(
+        return putRequest(
             createAction(UPDATE_ROOM_BOOKING),
             createAction(ROOM_BOOKING_UPDATED),
-            `${window.API_BASE_URL}/api/v1/summits/${currentSummit.id}/room-bookings/${entity.id}`,
+            `${window.API_BASE_URL}/api/v1/summits/${currentSummit.id}/locations/bookable-rooms/${entity.room_id}/reservations/${entity.id}`,
             normalizedEntity,
             authErrorHandler,
             entity
@@ -189,14 +186,14 @@ export const saveRoomBooking = (entity) => async (dispatch, getState) => {
     } else {
         const success_message = {
             title: T.translate("general.done"),
-            html: T.translate("edit_room_booking.room_booking_created"),
+            html: T.translate("edit_room_booking.offline_room_booking_created"),
             type: 'success'
         };
 
-        postRequest(
+        return postRequest(
             createAction(UPDATE_ROOM_BOOKING),
             createAction(ROOM_BOOKING_ADDED),
-            `${window.API_BASE_URL}/api/v1/summits/${currentSummit.id}/room-bookings`,
+            `${window.API_BASE_URL}/api/v1/summits/${currentSummit.id}/locations/bookable-rooms/${entity.room_id}/reservations/offline`,
             normalizedEntity,
             authErrorHandler,
             entity
@@ -437,7 +434,7 @@ export const deleteRoomBookingAttribute = (attributeTypeId, attributeValueId) =>
     );
 };
 
-export const getOfflineBookingRoomAvailability = (roomId, day) => async (dispatch, getState) => {
+export const getBookingRoomAvailability = (roomId, day) => async (dispatch, getState) => {
     const { currentSummitState } = getState();
     const accessToken = await getAccessTokenSafely();
     const { currentSummit } = currentSummitState;
@@ -450,7 +447,7 @@ export const getOfflineBookingRoomAvailability = (roomId, day) => async (dispatc
 
     return getRequest(
         null,
-        createAction(RECEIVE_OFFLINE_ROOM_BOOKING_AVAILABILITY),
+        createAction(RECEIVE_ROOM_BOOKING_AVAILABILITY),
         `${window.API_BASE_URL}/api/v1/summits/${currentSummit.id}/locations/bookable-rooms/${roomId}/availability/${day}`,
         (err, res) => customErrorHandler(err, res, { roomId, day }),
     )(params)(dispatch)
@@ -458,68 +455,6 @@ export const getOfflineBookingRoomAvailability = (roomId, day) => async (dispatc
             dispatch(stopLoading());
         });
 }
-
-export const saveOfflineRoomBooking = (entity) => async (dispatch, getState) => {
-    const { currentSummitState } = getState();
-    const accessToken = await getAccessTokenSafely();
-    const { currentSummit } = currentSummitState;
-
-    const params = {
-        access_token: accessToken,
-    };
-
-    dispatch(startLoading());
-
-    const normalizedEntity = normalizeEntity(entity);
-
-    const success_message = {
-        title: T.translate("general.done"),
-        html: T.translate("offline_room_booking.offline_room_booking_created"),
-        type: 'success'
-    };
-
-    return postRequest(
-        null,
-        createAction(OFFLINE_ROOM_BOOKING_ADDED),
-        `${window.API_BASE_URL}/api/v1/summits/${currentSummit.id}/locations/bookable-rooms/${entity.room_id}/reservations/offline`,
-        normalizedEntity,
-        authErrorHandler,
-        entity
-    )(params)(dispatch)
-        .then((payload) => {
-            dispatch(showMessage(
-                success_message,
-                () => { history.push(`/app/summits/${currentSummit.id}/room-bookings/${payload.response.id}`) }
-            ));
-        });
-}
-
-export const updateRoomBookingReservation = (entity) => async (dispatch, getState) => {
-    const { currentSummitState } = getState();
-    const accessToken = await getAccessTokenSafely();
-    const { currentSummit } = currentSummitState;
-
-    const params = {
-        access_token: accessToken,
-    };    
-
-    dispatch(startLoading());
-
-    const normalizedEntity = normalizeEntity(entity);
-
-    return putRequest(
-        createAction(UPDATE_ROOM_BOOKING_RESERVATION),
-        createAction(ROOM_BOOKING_RESERVATION_UPDATED),
-        `${window.API_BASE_URL}/api/v1/summits/${currentSummit.id}/locations/bookable-rooms/${entity.room_id}/reservations/${entity.id}`,
-        normalizedEntity,
-        authErrorHandler,
-        entity
-    )(params)(dispatch)
-        .then((payload) => {
-            dispatch(showSuccessMessage(T.translate("edit_room_booking.room_booking_saved")));
-        });
-}
-
 
 export const customErrorHandler = (err, res, { roomId, day }) => (dispatch, state) => {
     const code = err.status;
