+++ src/components/forms/room-booking-form.js
@@ -11,49 +11,69 @@
  * limitations under the License.
  **/
 
-import React from 'react'
+import React, { useEffect, useState } from 'react'
 import T from 'i18n-react/dist/i18n-react'
 import 'awesome-bootstrap-checkbox/awesome-bootstrap-checkbox.css'
-import {epochToMomentTimeZone} from 'openstack-uicore-foundation/lib/utils/methods'
+import { epochToMomentTimeZone } from 'openstack-uicore-foundation/lib/utils/methods'
 import { Input, Dropdown, MemberInput, DateTimePicker } from 'openstack-uicore-foundation/lib/components'
-import {isEmpty, scrollToError, shallowEqual} from "../../utils/methods";
+import { getAvailableBookingDates, getDayFromReservation, isEmpty, scrollToError, shallowEqual } from "../../utils/methods";
 
+import '../../styles/booking-room.less';
 
-class RoomBookingForm extends React.Component {
-    constructor(props) {
-        super(props);
+const RoomBookingForm = ({ history, entity, currentSummit, errors, availableSlots, onSubmit, getAvailableSlots }) => {
+    const [stateEntity, setStateEntity] = useState({ ...entity });
+    const [currentRoom, setCurrentRoom] = useState(null);
+    const [bookingDate, setBookingDate] = useState(null);
+    const [timeSlot, setTimeSlot] = useState(null);
 
-        this.state = {
-            entity: {...props.entity},
-            errors: props.errors
-        };
+    useEffect(() => {
+        if (entity.id) {
+            handleEntityChange(entity)
+        }
+    }, [])
 
-        this.handleChange = this.handleChange.bind(this);
-        this.handleSubmit = this.handleSubmit.bind(this);
+    useEffect(() => {
+        if (entity.id) {
+            handleEntityChange(entity)
+        } else {
+            setStateEntity({ ...entity })
+        }
+    }, [entity]);
+
+    const handleEntityChange = (newEntity) => {
+        const currentRoom = currentSummit.locations.find(l => l.id === newEntity.room_id);
+        const availableDates = getAvailableBookingDates(currentSummit);
+        const bookingDate = getDayFromReservation(newEntity, availableDates);
+        getAvailableSlots(currentRoom.id, bookingDate);
+        setStateEntity({ ...newEntity })
+        setCurrentRoom(currentRoom)
+        setBookingDate(bookingDate);
+        setTimeSlot(newEntity.start_datetime);
     }
 
-    componentDidUpdate(prevProps, prevState, snapshot) {
-        const state = {};
-        scrollToError(this.props.errors);
-
-        if(!shallowEqual(prevProps.entity, this.props.entity)) {
-            state.entity = {...this.props.entity};
-            state.errors = {};
-        }
-
-        if (!shallowEqual(prevProps.errors, this.props.errors)) {
-            state.errors = {...this.props.errors};
-        }
-
-        if (!isEmpty(state)) {
-            this.setState({...this.state, ...state})
-        }
+    const handleRoomChange = (ev) => {
+        let { value, id } = ev.target;
+        const currentRoom = currentSummit.locations.find(l => l.id === parseInt(value));
+        setCurrentRoom(currentRoom);
     }
 
-    handleChange(ev) {
-        let entity = {...this.state.entity};
-        let errors = {...this.state.errors};
-        let {value, id} = ev.target;
+    const handleBookingDateChange = (ev) => {
+        let { value } = ev.target;
+
+        getAvailableSlots(currentRoom.id, value);
+
+        setBookingDate(value);
+        setTimeSlot(null);
+    }
+
+    const handleTimeSlotChange = (ev) => {
+        let { value } = ev.target;
+        setTimeSlot(value);
+    }
+
+    const handleChange = (ev) => {
+        let entity = { ...stateEntity };
+        let { value, id } = ev.target;
 
         if (ev.target.type === 'number') {
             value = parseInt(ev.target.value);
@@ -63,147 +83,161 @@ class RoomBookingForm extends React.Component {
             value = ev.target.checked;
         }
 
-        errors[id] = '';
         entity[id] = value;
-        this.setState({entity: entity, errors: errors});
+        setStateEntity(entity)
     }
 
-    handleSubmit(ev) {
-        let entity = {...this.state.entity};
-        let {locationId} = this.props;
-
+    const handleSubmit = (ev) => {
         ev.preventDefault();
 
-        this.props.onSubmit(locationId, entity);
+        const { start_date, end_date } = availableSlots.find(e => e.start_date === timeSlot);
+
+        let normalizedEntity = {
+            id: stateEntity.id || null,
+            room_id: currentRoom.id,
+            start_datetime: start_date,
+            end_datetime: end_date,
+            owner_id: stateEntity.owner?.id,
+            currency: currentRoom.currency,
+            amount: currentRoom.time_slot_cost
+        }
+
+        onSubmit(normalizedEntity);
     }
 
-    hasErrors(field) {
-        let {errors} = this.state;
-        if(field in errors) {
+    const hasErrors = (field) => {
+        if (field in errors) {
             return errors[field];
         }
 
         return '';
     }
 
-    render() {
-        const {entity} = this.state;
-        const {currentSummit} = this.props;
+    let rooms_ddl = currentSummit.locations.filter(v => (v.class_name === 'SummitBookableVenueRoom')).map(l => {
+        return { label: l.name, value: l.id };
+    });
 
-        let rooms_ddl = currentSummit.locations.filter(v => (v.class_name === 'SummitBookableVenueRoom')).map(l => {
-            return {label: l.name, value: l.id};
-        });
+    const available_booking_dates_ddl = getAvailableBookingDates(currentSummit).map(v => ({ value: v.epoch, label: v.str }));
 
-        let currency_ddl = [
-            {label: 'USD', value: 'USD'},
-            {label: 'EUR', value: 'EUR'},
-        ];
+    const available_slots_ddl = availableSlots?.map(as => (
+        {
+            value: as.start_date,
+            label: `${epochToMomentTimeZone(as.start_date, currentSummit.time_zone_id).format('h:mm a')} - 
+                    ${epochToMomentTimeZone(as.end_date, currentSummit.time_zone_id).format('h:mm a')}
+                    ${as.is_free ? '' : ' - Booked'}`,
+            isDisabled: !as.is_free
+        }
+    ));
 
-        let status_ddl = [
-            {label: 'Reserved', value: 'Reserved'},
-            {label: 'Payed', value: 'Payed'},
-            {label: 'Requested Refund', value: 'RequestedRefund'},
-            {label: 'Refunded', value: 'Refunded'},
-            {label: 'Canceled', value: 'Canceled'}
-        ];
-
-        return (
-            <form className="room-booking-form">
-                <input type="hidden" id="id" value={entity.id} />
-                <div className="row form-group">
-                    <div className="col-md-4">
-                        <label> {T.translate("edit_room_booking.start")} *</label>
-                        <DateTimePicker
-                            id="start_datetime"
-                            onChange={this.handleChange}
-                            format={{date:"YYYY-MM-DD", time: "HH:mm"}}
-                            timezone={currentSummit.time_zone_id}
-                            value={epochToMomentTimeZone(entity.start_datetime, currentSummit.time_zone_id)}
-                        />
-                    </div>
-                    <div className="col-md-4">
-                        <label> {T.translate("edit_room_booking.end")} *</label>
-                        <DateTimePicker
-                            id="end_datetime"
-                            onChange={this.handleChange}
-                            format={{date:"YYYY-MM-DD", time: "HH:mm"}}
-                            timezone={currentSummit.time_zone_id}
-                            value={epochToMomentTimeZone(entity.end_datetime, currentSummit.time_zone_id)}
-                        />
-                    </div>
-                    <div className="col-md-4">
-                        <label> {T.translate("edit_room_booking.room")} </label>
-                        <Dropdown
-                            id="room_id"
-                            value={entity.room_id}
-                            options={rooms_ddl}
-                            placeholder={T.translate("edit_room_booking.placeholders.select_room")}
-                            onChange={this.handleChange}
-                            error={this.hasErrors('room_id')}
-                        />
-                    </div>
+    return (
+        <form className="room-booking-form">
+            <input type="hidden" id="id" value={stateEntity.id} />
+            <div className="row form-group">
+                <div className="col-md-4">
+                    <label> {T.translate("edit_room_booking.room")} </label>
+                    <Dropdown
+                        id="room_id"
+                        value={stateEntity.room_id}
+                        options={rooms_ddl}
+                        placeholder={T.translate("edit_room_booking.placeholders.select_room")}
+                        onChange={handleRoomChange}
+                        error={hasErrors('room_id')}
+                        isDisabled={stateEntity.id}
+                    />
                 </div>
-                <div className="row form-group">
-                    <div className="col-md-4">
-                        <label> {T.translate("edit_room_booking.status")}</label>
-                        <Dropdown
-                            id="status"
-                            value={entity.status}
-                            placeholder={T.translate("edit_room_booking.placeholders.select_status")}
-                            options={status_ddl}
-                            onChange={this.handleChange}
-                        />
+            </div>
+            {currentRoom?.id &&
+                <>
+                    <div className="row form-group">
+                        <div className="col-md-12">
+                            <div className="meeting-room">
+                                <div className="meeting-room-image">
+                                    <img src={currentRoom.image?.url} />
+                                </div>
+                                <div className="meeting-room-body">
+                                    <div className="meeting-room-header row">
+                                        <div className="col-xs-10">
+                                            <div className="meeting-room-title">{currentRoom.name}</div>
+                                        </div>
+                                    </div>
+                                    <div className="row">
+                                        <div className="meeting-room-info col-xs-12">
+                                            {T.translate("edit_room_booking.cost")}: {`${currentRoom.currency} ${currentRoom.time_slot_cost}`}
+                                        </div>
+                                    </div>
+                                    <div className="row">
+                                        <div className="meeting-room-info col-xs-6">
+                                            {T.translate("edit_room_booking.capacity")}: {currentRoom.capacity}
+                                        </div>
+                                    </div>
+                                    <div className="row">
+                                        <div className="meeting-room-amenities col-xs-12">
+                                            {currentRoom.attributes.length > 0 ? currentRoom.attributes.map(a => `${a.type.type}: ${a.value}`).join(' | ') : ''}
+                                        </div>
+                                    </div>
+                                    {currentRoom.floor &&
+                                        <div className="row">
+                                            <div className="meeting-room-info col-xs-12">
+                                                {T.translate("edit_room_booking.floor")}: {currentRoom.floor?.name}
+                                            </div>
+                                        </div>
+                                    }
+                                </div>
+                            </div>
+                        </div>
                     </div>
-                    <div className="col-md-4">
-                        <label> {T.translate("edit_room_booking.owner")}</label>
-                        <MemberInput
-                            id="owner"
-                            value={entity.owner}
-                            onChange={this.handleChange}
-                            getOptionLabel={
-                                (member) => {
-                                    return member.hasOwnProperty("email") ?
-                                        `${member.first_name} ${member.last_name} (${member.email})`:
-                                        `${member.first_name} ${member.last_name} (${member.id})`;
+                    <div className="row form-group">
+                        <div className="col-md-4">
+                            <label> {T.translate("edit_room_booking.date")}</label>
+                            <Dropdown
+                                id="booking_dates"
+                                value={bookingDate}
+                                placeholder={T.translate("edit_room_booking.placeholders.select_date")}
+                                options={available_booking_dates_ddl}
+                                onChange={handleBookingDateChange}
+                            />
+                        </div>
+                        {bookingDate &&
+                            <div className="col-md-4">
+                                <label> {T.translate("edit_room_booking.available_slots")}</label>
+                                <Dropdown
+                                    id="available_slots"
+                                    value={timeSlot}
+                                    placeholder={T.translate("edit_room_booking.placeholders.select_status")}
+                                    options={available_slots_ddl}
+                                    onChange={handleTimeSlotChange}
+                                />
+                            </div>
+                        }
+                        <div className="col-md-4">
+                            <label> {T.translate("edit_room_booking.owner")}</label>
+                            <MemberInput
+                                id="owner"
+                                value={stateEntity.owner}
+                                onChange={handleChange}
+                                getOptionLabel={
+                                    (member) => {
+                                        return member.hasOwnProperty("email") ?
+                                            `${member.first_name} ${member.last_name} (${member.email})` :
+                                            `${member.first_name} ${member.last_name} (${member.id})`;
+                                    }
                                 }
-                            }
-                        />
+                                isDisabled={stateEntity.id}
+                            />
+                        </div>
                     </div>
-                </div>
-                <div className="row form-group">
-                    <div className="col-md-4">
-                        <label> {T.translate("edit_room_booking.amount")}</label>
-                        <Input
-                            id="amount"
-                            type="number"
-                            value={entity.amount}
-                            onChange={this.handleChange}
-                            className="form-control"
-                            error={this.hasErrors('amount')}
-                        />
-                    </div>
-                    <div className="col-md-4">
-                        <label> {T.translate("edit_room_booking.status")}</label>
-                        <Dropdown
-                            id="currency"
-                            value={entity.currency}
-                            placeholder={T.translate("edit_room_booking.placeholders.select_currency")}
-                            options={currency_ddl}
-                            onChange={this.handleChange}
-                        />
-                    </div>
-                </div>
+                </>
+            }
 
 
-                <div className="row">
-                    <div className="col-md-12 submit-buttons">
-                        <input type="button" onClick={this.handleSubmit}
-                               className="btn btn-primary pull-right" value={T.translate("general.save")}/>
-                    </div>
+            <div className="row">
+                <div className="col-md-12 submit-buttons">
+                    <input type="button" onClick={handleSubmit}
+                        className="btn btn-primary pull-right" value={T.translate("general.save")} />
                 </div>
-            </form>
-        );
-    }
+            </div>
+        </form>
+    );
 }
 
 export default RoomBookingForm;
