+++ src/i18n/en.json
@@ -1317,22 +1317,7 @@
     "owner": "Owner",
     "status": "Status",
     "currency": "Currency",
-    "amount": "Amount",
-    "placeholders": {
-      "select_status": "Select Status",
-      "select_currency": "Select Currency",
-      "select_room": "Select Room"
-    }
-  },
-  "offline_room_booking": {
-    "room_booking": "Offline Room Booking",
-    "room": "Room",
-    "start": "Start",
-    "end": "End",
-    "owner": "Owner",
-    "status": "Status",
-    "currency": "Currency",
-    "amount": "Amount",
+    "amount": "Amount",    
     "date": "Date",
     "available_slots": "Available Slots",
     "cost": "Cost",
@@ -1345,7 +1330,7 @@
       "select_room": "Select Room",
       "select_date": "Select Date"
     }
-  },
+  },  
   "question_form": {
     "question_type": "Question Type",
     "question_id": "Question Identifier",
