+++ src/layouts/room-bookings-layout.js
@@ -20,7 +20,6 @@ import Restrict from '../routes/restrict';
 import RoomBookingListPage from '../pages/room_bookings/room-booking-list-page'
 import EditRoomBookingPage from '../pages/room_bookings/edit-room-booking-page'
 import NoMatchPage from "../pages/no-match-page";
-import OfflineRoomBookingPage from '../pages/room_bookings/offline-room-booking-page';
 
 
 class RoomBookingsLayout extends React.Component {
@@ -34,7 +33,7 @@ class RoomBookingsLayout extends React.Component {
                 <Switch>
                     <Route strict exact path={match.url} component={RoomBookingListPage}/>
                     <Route strict exact path={`${match.url}/:room_booking_id(\\d+)`} component={EditRoomBookingPage}/>
-                    <Route strict exact path={`${match.url}/new`} component={OfflineRoomBookingPage}/>
+                    <Route strict exact path={`${match.url}/new`} component={EditRoomBookingPage}/>
                     <Route component={NoMatchPage}/>
                 </Switch>
             </div>
