+++ src/styles/booking-room.less
